+++ src/tools.ts
@@ -1,5 +1,12 @@
 import maplibregl from "maplibre-gl";
+import type {
+  RequestParameters,
+  ResourceType,
+  RequestTransformFunction,
+} from "maplibre-gl";
 import { defaults } from "./defaults";
+import { config } from "./config";
+import { MAPTILER_SESSION_ID } from "./config";
 
 export function enableRTL() {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -42,4 +49,71 @@ export function DOMremove(node: HTMLElement) {
   if (node.parentNode) {
     node.parentNode.removeChild(node);
   }
+}
+
+/**
+ * This function is meant to be used as transformRequest by any Map instance created.
+ * It adds the session ID as well as the MapTiler Cloud key from the config to all the requests
+ * performed on MapTiler Cloud servers.
+ * @param url
+ * @param resourceType
+ * @returns
+ */
+export function maptilerCloudTransformRequest(
+  url: string,
+  resourceType?: ResourceType
+): RequestParameters {
+  let reqUrl = null;
+
+  try {
+    // The URL is expected to be absolute.
+    // Yet, if it's local we just return it without assuming a 'base' url (in the URL constructor)
+    // and we let the URL be locally resolved with a potential base path.
+    reqUrl = new URL(url);
+  } catch (e) {
+    return {
+      url,
+    };
+  }
+
+  if (reqUrl.host === defaults.maptilerApiHost) {
+    if (!reqUrl.searchParams.has("key")) {
+      reqUrl.searchParams.append("key", config.apiKey);
+    }
+
+    if (config.session) {
+      reqUrl.searchParams.append("mtsid", MAPTILER_SESSION_ID);
+    }
+  }
+
+  return {
+    url: reqUrl.href,
+  };
+}
+
+/**
+ * This combines a user-defined tranformRequest function (optionnal)
+ * with the MapTiler Cloud-specific one: maptilerCloudTransformRequest
+ * @param userDefinedRTF
+ * @returns
+ */
+export function combineTransformRequest(
+  userDefinedRTF: RequestTransformFunction = null
+): RequestTransformFunction {
+  return function (
+    url: string,
+    resourceType?: ResourceType
+  ): RequestParameters {
+    if (userDefinedRTF) {
+      const rp = userDefinedRTF(url, resourceType);
+      const rp2 = maptilerCloudTransformRequest(rp.url);
+
+      return {
+        ...rp,
+        ...rp2,
+      };
+    } else {
+      return maptilerCloudTransformRequest(url);
+    }
+  };
 }